requirements.txt
pydantic>=2.0.0
python-dotenv
openai
google-generativeai
requests
